Adds the ability to use SQS FIFO queues.
All SQS FIFO queues have the `.fifo` suffix at the end of their URL - this is how we determine if we are using a FIFO queue or not. The `MessageGroupId` (A required field for FIFO) is set using the key & id of the event's message.

Note: This commit does not add the the `MessageDeduplicationID` so the FIFO queue must have content-based deduplication enabled.